+++ packages/core/src/editor/index.tsx
@@ -1,4 +1,1 @@
 export * from './Editor';
-export * from './EditorStore';
-export * from './query';
-export * from './NodeHelpers';
