+++ packages/core/src/index.tsx
@@ -4,5 +4,6 @@ export * from './interfaces';
 export * from './hooks';
 export * from './editor';
 export * from './events';
+export * from './store';
 export * from './utils/testHelpers';
 export { ROOT_NODE } from '@craftjs/utils';
