+++ packages/core/src/store/index.tsx
@@ -1,0 +1,4 @@
+export * from './EditorStore';
+export * from './EditorStoreImpl';
+export * from './query';
+export * from './NodeHelpers';
